Refactor activate_ruleset with strategy pattern
